CHET-107: Move AWS Credentials web DTO into inner class
